add mcts_helper module
