+++ app/app.py
@@ -7,15 +7,12 @@ import networkx as nx
 # chrono imports
 import pychrono as chrono
 import rule_extention
-from control_optimisation import (create_grab_criterion_fun, create_traj_fun,
-                                  get_object_to_grasp)
+from control_optimisation import (create_grab_criterion_fun, create_traj_fun, get_object_to_grasp)
 
-import rostok.graph_generators.graph_environment as env
-from rostok.criterion.flags_simualtions import (FlagMaxTime, FlagNotContact,
-                                                FlagSlipout)
+from rostok.graph_generators.mcts_helper import prepare_mcts_state_and_helper, make_mcts_step
+from rostok.criterion.flags_simualtions import (FlagMaxTime, FlagNotContact, FlagSlipout)
 from rostok.graph_grammar.node import GraphGrammar
-from rostok.trajectory_optimizer.control_optimizer import (
-    ConfigRewardFunction, ControlOptimizer)
+from rostok.trajectory_optimizer.control_optimizer import (ConfigRewardFunction, ControlOptimizer)
 
 
 def plot_graph(graph: GraphGrammar):
@@ -25,6 +22,7 @@ def plot_graph(graph: GraphGrammar):
                      node_size=800,
                      labels={n: graph.nodes[n]["Node"].label for n in graph})
     plt.show()
+
 
 # %% Create extension rule vocabulary
 rule_vocabul, node_features = rule_extention.init_extension_rules()
@@ -63,23 +61,21 @@ finish = False
 G = GraphGrammar()
 max_numbers_rules = 2
 # Create graph environments for algorithm (not gym)
-graph_env = env.prepare_mcts_state_and_helper(G, rule_vocabul, control_optimizer, max_numbers_rules)
+graph_env = prepare_mcts_state_and_helper(G, rule_vocabul, control_optimizer, max_numbers_rules, Path("./results"))
 mcts_helper = graph_env.helper
 n_steps = 0
 #%% Run first algorithm
 while not finish:
-    finish, graph_env = env.make_mcts_step(searcher, graph_env, n_steps)
+    finish, graph_env = make_mcts_step(searcher, graph_env, n_steps)
     n_steps += 1
-    print(
-        f"number iteration: {iter}, counter actions: {graph_env.counter_action}"+ 
-        f"reward: {mcts_helper.report.get_best_info()[1]}"
-        )
+    print(f"number iteration: {n_steps}, counter actions: {graph_env.counter_action}" +
+          f"reward: {mcts_helper.report.get_best_info()[1]}")
 
 report = mcts_helper.report
 report.draw_best_graph()
 best_graph, reward, best_control = mcts_helper.report.get_best_info()
 func_reward = control_optimizer.create_reward_function(best_graph)
-res = - func_reward(best_control)
+res = -func_reward(best_control)
 print(res)
 report.plot_means()
 report.make_time_dependent_path()
